Fix C# implementations of cast_to_uint/1.
library/uint8.m:
library/uint16.m:
library/uint32.m:
library/uint64.m:
    Deal with the XXXs in the foreign_procs for cast_to_uint/1.

+++ library/uint16.m
@@ -454,8 +454,7 @@ cast_to_int(_) = _ :-
     cast_to_uint(U16::in) = (U::out),
     [will_not_call_mercury, promise_pure, thread_safe],
 "
-    // XXX This is probably wrong.
-    U = U16;
+    U = (uint) U16;
 ").
 
 :- pragma foreign_proc("Java",
