+++ library/uint64.m
@@ -450,8 +450,7 @@ cast_to_int(_) = _ :-
     cast_to_uint(U64::in) = (U::out),
     [will_not_call_mercury, promise_pure, thread_safe],
 "
-    // XXX This is probably wrong.
-    U = U64;
+    U = (uint) U64;
 ").
 
 :- pragma foreign_proc("Java",
